Split JSON Schema evaluator into a separate standalone module
