+++ .github/workflows/website-build.yml
@@ -19,6 +19,7 @@ jobs:
           -DJSONTOOLKIT_URI:BOOL=OFF
           -DJSONTOOLKIT_JSON:BOOL=OFF
           -DJSONTOOLKIT_JSONSCHEMA:BOOL=OFF
+          -DJSONTOOLKIT_EVALUATOR:BOOL=OFF
           -DJSONTOOLKIT_JSONPOINTER:BOOL=OFF
           -DJSONTOOLKIT_TESTS:BOOL=OFF
           -DJSONTOOLKIT_DOCS:BOOL=ON
