+++ .github/workflows/website-deploy.yml
@@ -29,6 +29,7 @@ jobs:
           -DJSONTOOLKIT_URI:BOOL=OFF
           -DJSONTOOLKIT_JSON:BOOL=OFF
           -DJSONTOOLKIT_JSONSCHEMA:BOOL=OFF
+          -DJSONTOOLKIT_EVALUATOR:BOOL=OFF
           -DJSONTOOLKIT_JSONPOINTER:BOOL=OFF
           -DJSONTOOLKIT_TESTS:BOOL=OFF
           -DJSONTOOLKIT_DOCS:BOOL=ON
