+++ CMakeLists.txt
@@ -9,6 +9,7 @@ include(vendor/noa/cmake/noa.cmake)
 option(JSONTOOLKIT_URI "Build the JSON Toolkit URI library" ON)
 option(JSONTOOLKIT_JSON "Build the JSON Toolkit JSON library" ON)
 option(JSONTOOLKIT_JSONSCHEMA "Build the JSON Toolkit JSON Schema library" ON)
+option(JSONTOOLKIT_EVALUATOR "Build the JSON Toolkit JSON Schema evaluator library" ON)
 option(JSONTOOLKIT_JSONPOINTER "Build the JSON Toolkit JSON Pointer library" ON)
 option(JSONTOOLKIT_JSONL "Build the JSON Toolkit JSONL library" ON)
 option(JSONTOOLKIT_TESTS "Build the JSON Toolkit tests" OFF)
@@ -48,7 +49,14 @@ if(JSONTOOLKIT_JSON AND JSONTOOLKIT_JSONPOINTER)
   add_subdirectory(src/jsonpointer)
 endif()
 
-if(JSONTOOLKIT_JSON AND JSONTOOLKIT_JSONSCHEMA)
+if(JSONTOOLKIT_URI AND JSONTOOLKIT_JSON AND
+   JSONTOOLKIT_JSONPOINTER AND JSONTOOLKIT_EVALUATOR)
+  add_subdirectory(src/evaluator)
+endif()
+
+if(JSONTOOLKIT_URI AND JSONTOOLKIT_JSON AND
+   JSONTOOLKIT_JSONPOINTER AND JSONTOOLKIT_EVALUATOR AND
+   JSONTOOLKIT_JSONSCHEMA)
   add_subdirectory(src/jsonschema)
 endif()
 
@@ -93,7 +101,14 @@ if(JSONTOOLKIT_TESTS)
     add_subdirectory(test/jsonpointer)
   endif()
 
-  if(JSONTOOLKIT_JSON AND JSONTOOLKIT_JSONSCHEMA)
+  if(JSONTOOLKIT_URI AND JSONTOOLKIT_JSON AND
+     JSONTOOLKIT_JSONPOINTER AND JSONTOOLKIT_EVALUATOR)
+    add_subdirectory(test/evaluator)
+  endif()
+
+  if(JSONTOOLKIT_URI AND JSONTOOLKIT_JSON AND
+     JSONTOOLKIT_JSONPOINTER AND JSONTOOLKIT_EVALUATOR AND
+     JSONTOOLKIT_JSONSCHEMA)
     add_subdirectory(test/jsonschema)
   endif()
