+++ config.cmake.in
@@ -9,6 +9,7 @@ if(NOT JSONTOOLKIT_COMPONENTS)
   list(APPEND JSONTOOLKIT_COMPONENTS jsonl)
   list(APPEND JSONTOOLKIT_COMPONENTS jsonpointer)
   list(APPEND JSONTOOLKIT_COMPONENTS jsonschema)
+  list(APPEND JSONTOOLKIT_COMPONENTS evaluator)
 endif()
 
 foreach(component ${JSONTOOLKIT_COMPONENTS})
@@ -35,6 +36,7 @@ foreach(component ${JSONTOOLKIT_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_uri.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_json.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_jsonpointer.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_evaluator.cmake")
 
     # GCC does not allow the use of std::promise, std::future
     # without compiling with pthreads support.
@@ -46,6 +48,10 @@ foreach(component ${JSONTOOLKIT_COMPONENTS})
     endif()
 
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_jsonschema.cmake")
+  elseif(component STREQUAL "evaluator")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_uri.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_json.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_jsontoolkit_evaluator.cmake")
   else()
     message(FATAL_ERROR "Unknown JSON Toolkit component: ${component}")
   endif()
