+++ doxygen/Doxyfile.in
@@ -2314,6 +2314,7 @@ PREDEFINED             += SOURCEMETA_JSONTOOLKIT_JSON_EXPORT=
 PREDEFINED             += SOURCEMETA_JSONTOOLKIT_JSONL_EXPORT=
 PREDEFINED             += SOURCEMETA_JSONTOOLKIT_JSONPOINTER_EXPORT=
 PREDEFINED             += SOURCEMETA_JSONTOOLKIT_JSONSCHEMA_EXPORT=
+PREDEFINED             += SOURCEMETA_JSONTOOLKIT_EVALUATOR_EXPORT=
 PREDEFINED             += DOXYGEN
 
 # If the MACRO_EXPANSION and EXPAND_ONLY_PREDEF tags are set to YES then this
@@ -2328,6 +2329,7 @@ EXPAND_AS_DEFINED      += SOURCEMETA_JSONTOOLKIT_JSON_EXPORT
 EXPAND_AS_DEFINED      += SOURCEMETA_JSONTOOLKIT_JSONL_EXPORT
 EXPAND_AS_DEFINED      += SOURCEMETA_JSONTOOLKIT_JSONPOINTER_EXPORT
 EXPAND_AS_DEFINED      += SOURCEMETA_JSONTOOLKIT_JSONSCHEMA_EXPORT
+EXPAND_AS_DEFINED      += SOURCEMETA_JSONTOOLKIT_EVALUATOR_EXPORT
 
 # If the SKIP_FUNCTION_MACROS tag is set to YES then doxygen's preprocessor will
 # remove all references to function-like macros that are alone on a line, have
