+++ src/evaluator/include/sourcemeta/jsontoolkit/evaluator.h
@@ -1,0 +1,147 @@
+#ifndef SOURCEMETA_JSONTOOLKIT_EVALUATOR_H_
+#define SOURCEMETA_JSONTOOLKIT_EVALUATOR_H_
+
+#include "evaluator_export.h"
+
+#include <sourcemeta/jsontoolkit/evaluator_error.h>
+#include <sourcemeta/jsontoolkit/evaluator_template.h>
+
+#include <sourcemeta/jsontoolkit/json.h>
+#include <sourcemeta/jsontoolkit/jsonpointer.h>
+
+#include <cstdint>    // std::uint8_t
+#include <functional> // std::function
+
+/// @defgroup evaluator Evaluator
+/// @brief A high-performance JSON Schema evaluator
+///
+/// This functionality is included as follows:
+///
+/// ```cpp
+/// #include <sourcemeta/jsontoolkit/evaluator.h>
+/// ```
+
+namespace sourcemeta::jsontoolkit {
+
+/// @ingroup evaluator
+/// Represents the mode of evalution
+enum class SchemaCompilerEvaluationMode : std::uint8_t {
+  /// Attempt to get to a boolean result as fast as possible
+  Fast,
+  /// Perform a full schema evaluation
+  Exhaustive
+};
+
+/// @ingroup evaluator
+/// Represents the state of a step evaluation
+enum class SchemaCompilerEvaluationType : std::uint8_t { Pre, Post };
+
+/// @ingroup evaluator
+/// A callback of this type is invoked after evaluating any keyword. The
+/// arguments go as follows:
+///
+/// - The stage at which the step in question is
+/// - Whether the evaluation was successful or not (always true before
+/// evaluation)
+/// - The step that was just evaluated
+/// - The evaluation path
+/// - The instance location
+/// - The annotation result, if any (otherwise null)
+///
+/// You can use this callback mechanism to implement arbitrary output formats.
+using SchemaCompilerEvaluationCallback =
+    std::function<void(const SchemaCompilerEvaluationType, bool,
+                       const SchemaCompilerTemplate::value_type &,
+                       const WeakPointer &, const WeakPointer &, const JSON &)>;
+
+// TODO: Support standard output formats too
+
+/// @ingroup evaluator
+///
+/// This function evaluates a schema compiler template in validation mode,
+/// returning a boolean without error information. For example:
+///
+/// ```cpp
+/// #include <sourcemeta/jsontoolkit/json.h>
+/// #include <sourcemeta/jsontoolkit/jsonschema.h>
+/// #include <sourcemeta/jsontoolkit/evaluator.h>
+/// #include <cassert>
+///
+/// const sourcemeta::jsontoolkit::JSON schema =
+///     sourcemeta::jsontoolkit::parse(R"JSON({
+///   "$schema": "https://json-schema.org/draft/2020-12/schema",
+///   "type": "string"
+/// })JSON");
+///
+/// const auto schema_template{sourcemeta::jsontoolkit::compile(
+///     schema, sourcemeta::jsontoolkit::default_schema_walker,
+///     sourcemeta::jsontoolkit::official_resolver,
+///     sourcemeta::jsontoolkit::default_schema_compiler)};
+///
+/// const sourcemeta::jsontoolkit::JSON instance{"foo bar"};
+/// const auto result{sourcemeta::jsontoolkit::evaluate(
+///   schema_template, instance)};
+/// assert(result);
+/// ```
+auto SOURCEMETA_JSONTOOLKIT_EVALUATOR_EXPORT
+evaluate(const SchemaCompilerTemplate &steps, const JSON &instance) -> bool;
+
+/// @ingroup evaluator
+///
+/// This function evaluates a schema compiler template, executing the given
+/// callback at every step of the way. For example:
+///
+/// ```cpp
+/// #include <sourcemeta/jsontoolkit/json.h>
+/// #include <sourcemeta/jsontoolkit/jsonschema.h>
+/// #include <sourcemeta/jsontoolkit/evaluator.h>
+/// #include <cassert>
+/// #include <iostream>
+///
+/// const sourcemeta::jsontoolkit::JSON schema =
+///     sourcemeta::jsontoolkit::parse(R"JSON({
+///   "$schema": "https://json-schema.org/draft/2020-12/schema",
+///   "type": "string"
+/// })JSON");
+///
+/// const auto schema_template{sourcemeta::jsontoolkit::compile(
+///     schema, sourcemeta::jsontoolkit::default_schema_walker,
+///     sourcemeta::jsontoolkit::official_resolver,
+///     sourcemeta::jsontoolkit::default_schema_compiler)};
+///
+/// static auto callback(
+///     bool result,
+///     const sourcemeta::jsontoolkit::SchemaCompilerTemplate::value_type &step,
+///     const sourcemeta::jsontoolkit::Pointer &evaluate_path,
+///     const sourcemeta::jsontoolkit::Pointer &instance_location,
+///     const sourcemeta::jsontoolkit::JSON &document,
+///     const sourcemeta::jsontoolkit::JSON &annotation) -> void {
+///   std::cout << "TYPE: " << (result ? "Success" : "Failure") << "\n";
+///   std::cout << "STEP:\n";
+///   sourcemeta::jsontoolkit::prettify(sourcemeta::jsontoolkit::to_json({step}),
+///                                     std::cout);
+///   std::cout << "\nEVALUATE PATH:";
+///   sourcemeta::jsontoolkit::stringify(evaluate_path, std::cout);
+///   std::cout << "\nINSTANCE LOCATION:";
+///   sourcemeta::jsontoolkit::stringify(instance_location, std::cout);
+///   std::cout << "\nANNOTATION:\n";
+///   sourcemeta::jsontoolkit::prettify(annotation, std::cout);
+///   std::cout << "\n";
+/// }
+///
+/// const sourcemeta::jsontoolkit::JSON instance{"foo bar"};
+/// const auto result{sourcemeta::jsontoolkit::evaluate(
+///   schema_template, instance,
+///   sourcemeta::jsontoolkit::SchemaCompilerEvaluationMode::Fast,
+///   callback)};
+///
+/// assert(result);
+/// ```
+auto SOURCEMETA_JSONTOOLKIT_EVALUATOR_EXPORT
+evaluate(const SchemaCompilerTemplate &steps, const JSON &instance,
+         const SchemaCompilerEvaluationMode mode,
+         const SchemaCompilerEvaluationCallback &callback) -> bool;
+
+} // namespace sourcemeta::jsontoolkit
+
+#endif
