+++ test/packaging/find_package/CMakeLists.txt
@@ -10,3 +10,4 @@ target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::json)
 target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::jsonschema)
 target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::jsonpointer)
 target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::jsonl)
+target_link_libraries(jsontoolkit_hello PRIVATE sourcemeta::jsontoolkit::evaluator)
